chore: 도메인 로직을 제외한 나머지 ts에서 js로 마이그레이션
